b2b/requirements.txt
Django>=4.1.7,<4.2
django-environ>=0.10.0,<0.11
psycopg2>=2.9.5,<2.10
djangorestframework>=3.14.0,<3.15.0
django-nested-admin>=4.0.2,<4.1
drf-nested-routers>=0.93.4,<0.94
redis>=4.5.1,<4.6
celery>=5.2.7,<5.3
django-templated-mail>=1.1.1,<1.2
whitenoise>=6.4.0,<6.5
gunicorn>=20.1.0,<20.2
django-cors-headers>=3.14.0,<3.15
